requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beanie==1.21.0
click==8.1.7
dnspython==2.4.2
fastapi==0.103.1
h11==0.14.0
httptools==0.6.0
idna==3.4
lazy-model==0.1.0b0
motor==3.3.1
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
